+++ apps/backend/src/csr-programme/csr-programme.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import { CSRProgrammeService } from './csr-programme.service';
 import { PrismaModule } from '../prisma/prisma.module';
+import { CSRProgrammeController } from './csr-programme.controller';
 
 @Module({
   imports: [PrismaModule],
+  controllers: [CSRProgrammeController],
   providers: [CSRProgrammeService],
   exports: [CSRProgrammeService],
 })